add documentation comments and do some cosmetic refactoring
